requirements.txt
absl-py==2.3.0
accelerate==1.7.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
aiosqlite==0.21.0
albucore==0.0.24
albumentations==2.0.8
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
astunparse==1.6.3
async-timeout==4.0.3
attrs==25.3.0
auto_gptq==0.7.1
av==14.4.0
banks==2.1.2
beautifulsoup4==4.13.4
bitsandbytes==0.46.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
dataclasses-json==0.6.7
datasets==3.6.0
decorator==5.2.1
Deprecated==1.2.18
diffusers==0.33.1
dill==0.3.8
dirtyjson==1.0.8
diskcache==5.6.3
distro==1.9.0
einops==0.8.1
evaluate==0.4.3
exceptiongroup==1.3.0
executing==2.2.0
fastapi==0.115.12
ffmpy==0.6.0
filelock==3.18.0
filetype==1.2.0
flatbuffers==25.2.10
fonttools==4.58.4
fpdf==1.7.2
frozenlist==1.7.0
fsspec==2025.3.0
gast==0.6.0
gekko==1.3.0
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
gradio==5.34.0
gradio_client==1.10.3
greenlet==3.2.3
griffe==1.7.3
groovy==0.1.2
grpcio==1.73.0
guidance==0.2.1
guidance-stitch==0.1.5
h11==0.16.0
h5py==3.14.0
hf-xet==1.1.4
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.0
humanfriendly==10.0
idna==3.10
importlib_metadata==8.7.0
ipython==8.37.0
ipywidgets==8.1.7
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyterlab_widgets==3.0.15
keras==3.10.0
kiwisolver==1.4.8
langchain==0.3.25
langchain-core==0.3.65
langchain-text-splitters==0.3.8
langsmith==0.3.45
libclang==18.1.1
llama-cloud==0.1.26
llama-cloud-services==0.6.34
llama-index==0.12.42
llama-index-agent-openai==0.4.11
llama-index-cli==0.4.3
llama-index-core==0.12.42
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.7.7
llama-index-llms-openai==0.4.7
llama-index-multi-modal-llms-openai==0.5.1
llama-index-program-openai==0.3.2
llama-index-question-gen-openai==0.3.1
llama-index-readers-file==0.4.9
llama-index-readers-llama-parse==0.4.0
llama-parse==0.6.34
llguidance==0.6.31
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_dtypes==0.5.1
mpmath==1.3.0
multidict==6.4.4
multiprocess==0.70.16
mypy_extensions==1.1.0
namex==0.1.0
narwhals==1.43.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==2.1.3
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
onnx==1.18.0
onnxruntime==1.22.0
openai==1.88.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
optimum==1.26.1
optree==0.16.0
ordered-set==4.1.0
orjson==3.10.18
packaging==24.2
pandas==2.2.3
parso==0.8.4
pdfrw==0.4
peft==0.15.2
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==5.29.5
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
pydicom==3.0.1
pydub==0.25.1
Pygments==2.19.1
PyMuPDF==1.26.3
pyparsing==3.2.3
pypdf==5.6.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
qwen-vl-utils==0.0.11
referencing==0.36.2
regex==2024.11.6
reportlab==4.4.2
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.0.0
rouge==1.0.1
rouge-score==0.1.2
rpds-py==0.25.1
ruff==0.11.13
safehttpx==0.1.6
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
seaborn==0.13.2
semantic-version==2.10.0
sentence-transformers==4.1.0
sentencepiece==0.2.0
shellingham==1.5.4
simsimd==6.4.9
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
stack-data==0.6.3
starlette==0.46.2
streamlit==1.45.1
stringzilla==3.12.5
striprtf==0.0.26
sympy==1.14.0
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.1.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
toml==0.10.2
tomlkit==0.13.3
torch==2.7.1
torchaudio==2.7.1
torchvision==0.22.1
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.52.4
triton==3.3.1
typer==0.16.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.3
watchdog==6.0.0
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.3
widgetsnbextension==4.0.14
wrapt==1.17.2
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
